use run-test util
